server/requirements.txt
PyMuPDF==1.22.5
python-docx==0.8.11
pdfplumber==0.10.2
transformers==4.30.2
torch==2.0.1 